requirements.txt
Flask==3.0.0
pyTelegramBotAPI==4.14.0
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
aiosqlite==0.19.0
python-dotenv==1.0.0
gunicorn==22.0.0
